Restored solution dependency graph display
